Move to integer grid, use optionals,  apply minor redesign svg logging, generate better graphs

+++ cgal/main.cpp
@@ -3,13 +3,16 @@
 #include <iomanip>
 #include <fstream>
 #include <vector>
-
-#include <CGAL/circulator.h>
+#include <set>
+#include <algorithm>
 
 #include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
 #include <CGAL/Delaunay_triangulation_2.h>
 #include <CGAL/Triangulation_face_base_with_info_2.h>
 #include <CGAL/point_generators_2.h>
+#include <CGAL/Random.h>
+
+#include "include/optional.hpp"
 
 typedef CGAL::Exact_predicates_inexact_constructions_kernel Kernel;
 typedef Kernel::Point_2 Point;
@@ -22,7 +25,6 @@ typedef CGAL::Delaunay_triangulation_2<Kernel, TriangulationStructure> DelaunayT
 typedef DelaunayTriangulation::Face_handle Face;
 typedef DelaunayTriangulation::Segment Segement;
 typedef DelaunayTriangulation::Edge Edge;
-typedef DelaunayTriangulation::Vertex_circulator Vertex_circulator;
 
 typedef CGAL::Random_points_in_square_2<DelaunayTriangulation::Point> RandomPoints;
 typedef CGAL::Counting_iterator<RandomPoints, DelaunayTriangulation::Point> PointsIterator;
@@ -30,16 +32,18 @@ typedef CGAL::Counting_iterator<RandomPoints, DelaunayTriangulation::Point> Poin
 #define PI 3.14159265358979323846
 
 using namespace std;
+using nonstd::nullopt;
+using nonstd::optional;
 
 static const string LOG_FOLDER("output");
-static double D(5.0);
+static int D(44000);
 
-class WalkLog
+class SVGLog
 {
   private:
     std::ofstream f;
 
-    const string OPEN = R"(
+    string OPEN = R"(
         <!DOCTYPE html>
             <html>
                 <head>
@@ -62,14 +66,14 @@ class WalkLog
                             justify-content: center;
                         }
                         .mesh-line {
-                            stroke-width: 0.01;
+                            stroke-width: 100;
                             stroke:rgb(200, 200, 200);
                         }
                         .path-face {
                             fill:rgba(100, 200, 100, 0.5);
                         }
                         #arrow {
-                            stroke-width: 0.01;
+                            stroke-width: 200;
                             stroke:rgb(0, 0, 200);
                         }
                         #arrow-head {
@@ -79,7 +83,9 @@ class WalkLog
                 </head>
                 <body>
                     <div id="mesh-div">
-                        <svg viewBox="-5 -5 10 10" preserveAspectRatio="xMidYMid meet" xmlns="http://www.w3.org/2000/svg" version="1.1" id="mesh-svg">
+                        <svg viewBox=")" +
+                  to_string(0 - D) + " " + to_string(0 - D) + " " + to_string(D + D) + " " + to_string(D + D) +
+                  R"(" preserveAspectRatio="xMidYMid meet" xmlns="http://www.w3.org/2000/svg" version="1.1" id="mesh-svg">
                             <defs>
                                 <marker id="arrow-head" markerWidth="10" markerHeight="10" refX="7" refY="3" orient="auto" markerUnits="strokeWidth">
                                     <path d="M0,0 L0,6 L9,3 z" class="arrowhead"/>
@@ -87,7 +93,7 @@ class WalkLog
                             </defs>
                             <g transform="scale(1, -1) ">
                             )";
-    const string CLOSE = R"(
+    string CLOSE = R"(
                             </g>
                             Sorry, your browser does not support inline SVG.
                         </svg>
@@ -96,11 +102,13 @@ class WalkLog
             </html>)";
 
   public:
-    WalkLog(const string &filename)
+    SVGLog(const string &filename)
     {
         f.open(filename, ios::trunc | ios::in);
         f << OPEN;
     }
+
+    SVGLog() {}
 
     void setup(const DelaunayTriangulation &T)
     {
@@ -130,7 +138,7 @@ class WalkLog
         }
     }
 
-    void log(const Face &face)
+    void face(const Face &face)
     {
         auto
             a = face->vertex(0)->point(),
@@ -145,7 +153,7 @@ class WalkLog
           << endl;
     }
 
-    void log(const Point &point, const string &color = "red", double r = 0.05)
+    void circle(const Point &point, const string &color = "red", double r = 1)
     {
         f << "<circle "
           << "cx=\"" << point.x() << "\" cy=\"" << point.y() << "\" r=\"" << r << "\""
@@ -153,11 +161,26 @@ class WalkLog
           << endl;
     }
 
-    void log(const Point &origin, const Point &target)
+    void arrow(const Point &origin, const Point &target)
     {
         f << "<line x1=\"" << origin.x() << "\" y1=\"" << origin.y() << "\""
           << " x2=\"" << target.x() << "\" y2=\"" << target.y() << "\""
-          << " id=\"arrow\" marker-end=\"url(#arrow-head)\"/>" << endl;
+          << " id=\"arrow\" marker-end=\"url(#arrow-head)\" />" << endl;
+    }
+
+    void line(const Point &origin, const Point &target)
+    {
+        f << "<line x1=\"" << origin.x() << "\" y1=\"" << origin.y() << "\""
+          << " x2=\"" << target.x() << "\" y2=\"" << target.y() << "\""
+          << " class=\"mesh-line\"/>" << endl;
+    }
+
+    void text(const Point &pos, const string &text, const string &style)
+    {
+        f << "<g transform=\"scale(1, -1)\">"
+          << "<text x=\"" << pos.x() << "\" y=\"" << 0 - pos.y() << "\" style=\""
+          << style << "\">" << text << "</text>"
+          << "</g>" << endl;
     }
 
     void close()
@@ -302,7 +325,7 @@ int celestial_walk(
     const Point &origin,
     const Point &target,
     const DelaunayTriangulation &T,
-    WalkLog &walk_log)
+    optional<SVGLog> &walk_log)
 {
     int tests(0);
     // Select an edge of the face that contains the origin;
@@ -324,31 +347,43 @@ int celestial_walk(
     // rim = rim.next();
     // cout << rim << endl;
 
-    walk_log.log(rim.origin(), rim.target());
+    if (walk_log)
+    {
+        walk_log->arrow(rim.origin(), rim.target());
+    }
 
     // Select an edge in the face such that the target point is on the left of it.
     if (++tests && strictlyRightOf(rim.origin(), rim.target(), target))
     {
         // Log the face we will leave.
-
-        walk_log.log(rim.face());
-
+        if (walk_log)
+        {
+            walk_log->face(rim.face());
+        }
         // Twin is assumed to exist.
         rim = rim.twin();
     }
 
-    walk_log.log(rim.origin(), rim.target());
-
+    if (walk_log)
+    {
+        walk_log->arrow(rim.origin(), rim.target());
+    }
     auto rim2 = rim.next();
 
-    walk_log.log(rim2.origin(), rim2.target());
+    if (walk_log)
+    {
+        walk_log->arrow(rim2.origin(), rim2.target());
+    }
 
     int n = 0;
     while (rim != rim2 && n < 10000)
     {
         n++;
 
-        walk_log.flush();
+        if (walk_log)
+        {
+            walk_log->flush();
+        }
 
         if (++tests && strictlyRightOf(rim2.origin(), rim2.target(), target))
         {
@@ -358,31 +393,49 @@ int celestial_walk(
             {
                 rim2 = rim2.next();
 
-                walk_log.log(rim2.origin(), rim2.target());
+                if (walk_log)
+                {
+                    walk_log->arrow(rim2.origin(), rim2.target());
+                }
 
                 approxBisector = computeApproxBisector(rim2);
                 obtuse = computeObtuseness(rim2);
             }
 
-            walk_log.log(rim2.face());
+            if (walk_log)
+            {
+                walk_log->face(rim2.face());
+            }
 
             // Twin is assumed to exist.
             rim = rim2.twin();
 
-            walk_log.log(rim.origin(), rim.target());
+            if (walk_log)
+            {
+                walk_log->arrow(rim.origin(), rim.target());
+            }
 
             rim2 = rim.next();
 
-            walk_log.log(rim2.origin(), rim2.target());
+            if (walk_log)
+            {
+                walk_log->arrow(rim2.origin(), rim2.target());
+            }
         }
         else
         {
             rim2 = rim2.next();
 
-            walk_log.log(rim2.origin(), rim2.target());
+            if (walk_log)
+            {
+                walk_log->arrow(rim2.origin(), rim2.target());
+            }
         }
 
-        walk_log.log(rim2.face());
+        if (walk_log)
+        {
+            walk_log->face(rim2.face());
+        }
     }
 
     return tests;
@@ -399,50 +452,63 @@ pair<double, double> mean_stdev(const vector<int> &v)
     return pair<double, double>(mean, sqrt(var));
 }
 
-int main()
+int main(int argc, char *argv[])
 {
+    if (argc != 5)
+    {
+        cout << "main {triangulations} {points in triangulation} {direction change} {write logs (0 or 1)}" << endl;
+        return 0;
+    }
     // Create the log folder if does not exist and remove all logs.
     string c("mkdir -p " + LOG_FOLDER + " && rm -rf ./" + LOG_FOLDER + "/*");
     system(c.c_str());
 
-    // Describe extrem point of the surface to triangulate.
-    vector<Point> extrema{Point(-D, -D), Point(-D, D), Point(D, -D), Point(D, D)};
-
+    // Number of triangulations.
+    int TRINAGULATIONS = atoi(argv[1]);
     // Number of points in trinagulation.
-    int POINTS = 1000;
-    int TRINAGULATIONS = 10;
-
+    int POINTS = atoi(argv[2]);
     // Every time the walking direction changes by step degrees.
-    int step = 45;
+    int step = atoi(argv[3]);
+    bool log_it = atoi(argv[4]) == 1;
+
     // There is a vector of stats for each angle.
-    vector<vector<int>> stats(360 / step, vector<int>());
+    vector<vector<int>> data(360 / step, vector<int>());
 
     // Always walk from the origin.
     Point origin(0.0, 0.0);
+    // Describe extrem point of the surface to triangulate.
+    set<Point> extrema{Point(-D, -D), Point(-D, D), Point(D, -D), Point(D, D)};
 
     int t = 0;
     while (t < TRINAGULATIONS)
     {
         // Generate a random Delaunay triangulation.
         DelaunayTriangulation DT;
-        RandomPoints rand_it(D);
-        PointsIterator first(rand_it);
-        PointsIterator last(rand_it, POINTS);
+        set<Point> points;
+        CGAL::Random rnd;
+        generate_n(inserter(points, points.begin()), POINTS, [&rnd]() {
+            // 1 - D, so the left border is not included.
+            return Point(rnd.get_int(1 - D, D), rnd.get_int(1 - D, D));
+        });
 
         DT.insert(extrema.begin(), extrema.end());
-        DT.insert(first, last);
+        DT.insert(points.begin(), points.end());
 
         // Iterate over compass directions.
         for (int angle = 0; angle < 360; angle += step)
         {
-            // Setup the logger.
-            WalkLog walk_log(LOG_FOLDER + "/triangulation_t(" + to_string(t) + ")_angle(" + to_string(angle) + ").html");
+            // SVGLog walk_log(LOG_FOLDER + "/triangulation_t(" + to_string(t) + ")_angle(" + to_string(angle) + ").html");
+            optional<SVGLog> walk_log =
+                log_it ? optional<SVGLog>(SVGLog(LOG_FOLDER + "/triangulation_t(" + to_string(t) + ")_angle(" + to_string(angle) + ").html")) : nullopt;
 
-            walk_log.setup(DT);
-            // Mark a reference point.
-            walk_log.log(Point(D, D), "green");
-            walk_log.flush();
-            // Logger setup is finished.
+            if (walk_log)
+            {
+                // Setup the logger.
+                walk_log->setup(DT);
+                // Mark a reference point.
+                walk_log->circle(Point(D, D), "green", 500);
+                walk_log->flush();
+            }
 
             // Fix direction.
             double r = 0.8 * D;
@@ -450,11 +516,14 @@ int main()
 
             // Perform the walk.
             int tests = celestial_walk(origin, target, DT, walk_log);
-            stats[angle / step].push_back(tests);
+            data[angle / step].push_back(tests);
 
-            walk_log.log(origin, "blue");
-            walk_log.log(target, "red");
-            walk_log.close();
+            if (walk_log)
+            {
+                walk_log->circle(origin, "blue", 500);
+                walk_log->circle(target, "red", 500);
+                walk_log->close();
+            }
         }
 
         cout << (t + 1) << "/" << TRINAGULATIONS << endl;
@@ -462,6 +531,52 @@ int main()
     }
 
     // Process stats.
+    struct Statistic
+    {
+        int angle;
+        double mean;
+        double std_dev;
+    };
+
+    vector<Statistic> stats;
+
+    auto angle = 0;
+    auto d = mean_stdev(data[0]);
+    double m(d.first), M(d.first);
+    stats.push_back((Statistic){angle, d.first, d.second});
+
+    for (auto i = 1; i < data.size(); ++i)
+    {
+        angle = step * i;
+        d = mean_stdev(data[i]);
+        stats.push_back((Statistic){angle, d.first, d.second});
+
+        if (d.first < m)
+        {
+            m = d.first;
+        }
+        if (d.first > M)
+        {
+            M = d.first;
+        }
+    }
+
+    // x-axis is going to be rescaled to angle values.
+    auto x_scaler = [](double a) {
+        // add a little space around by add 30 degree to ech side of [0, 360).
+        // return 2 * D * (a - (-30)) / ((360 + 30) - (-30)) - D;
+        return 2 * D * (a + 30) / 390 - D;
+    };
+    // y-axis is going to be rescaled to fit mean value.
+    auto y_scaler = [&m, &M](double v) {
+        // add a little space around by multiplying m and M.
+        return 2 * D * (v - 0.9 * m) / (1.1 * M - 0.9 * m) - D;
+    };
+
+    SVGLog graph(LOG_FOLDER + "/graph.html");
+    graph.text(Point(-D, y_scaler(m)), to_string(m), "font: bold 2000px sans-serif;");
+    graph.text(Point(-D, y_scaler(M)), to_string(M), "font: bold 2000px sans-serif;");
+
     cout << endl
          << endl
          << setw(10) << "Angle"
@@ -469,12 +584,25 @@ int main()
          << " | " << setw(10) << "Std.Dev." << endl;
     for (auto i = 0; i < stats.size(); i++)
     {
-        auto dist = mean_stdev(stats[i]);
-        cout << setw(10) << step * i
-             << " | " << setw(10) << dist.first
-             << " | " << setw(10) << dist.second
+        cout << setw(10) << stats[i].angle
+             << " | " << setw(10) << stats[i].mean
+             << " | " << setw(10) << stats[i].std_dev
              << endl;
+
+        graph.line(
+            Point(x_scaler(stats[i].angle), -D),
+            Point(x_scaler(stats[i].angle), D));
+        if (i % 2 == 0)
+        {
+            graph.text(
+                Point(x_scaler(stats[i].angle), -D + 2000),
+                to_string(stats[i].angle), "font: bold 2000px sans-serif;");
+        }
+
+        graph.circle(Point(x_scaler(stats[i].angle), y_scaler(stats[i].mean)), "green", 500);
     }
+
+    graph.close();
 
     return 0;
 }
